Added support in C++ for: integer down-casts, verification of .hpp files, inclusion of C++ standard header files (that don't have file extensions). (#493)

+++ tests/cxx/casts.cpp
@@ -1,9 +1,9 @@
-unsigned long long alignmentRest(void const * const pointer, unsigned int const byte_count)
+unsigned int alignmentRest(void const * const pointer, unsigned int const byte_count)
 //@ requires byte_count > 0;
 //@ ensures result <= byte_count;
 {
 	unsigned long long const p = (unsigned long long) pointer;
 	//@ div_rem_nonneg(p, byte_count);
 
-	return byte_count - (p % byte_count);
+	return (unsigned int) (byte_count - (p % byte_count));
 }
